Move BookReaderTextFragment to TextSelectionManager
